消除警告

+++ src/QCustomUi/QCtmMultiPageStringListModel.cpp
@@ -97,7 +97,8 @@ const QStringList& QCtmMultiPageStringListModel::stringList() const { return m_i
 */
 int QCtmMultiPageStringListModel::rowCount(const QModelIndex& parent /* = QModelIndex() */) const
 {
-    return std::min<size_t>(std::max<size_t>(m_impl->strings.size() - currentPage() * pageRowCount(), 0LL), pageRowCount());
+    return static_cast<int>(
+        std::min<size_t>(std::max<size_t>(m_impl->strings.size() - currentPage() * pageRowCount(), 0LL), pageRowCount()));
 }
 
 /*!
